Refactor terraswap tlvs (#2)
