+++ src/components/Dashboard/fetchAprData.js
@@ -1,6 +1,6 @@
 import dayjs from 'dayjs'
 
-export const fetchAprData = (report) => {
+export const fetchData = (report, widgetFormatter, fieldKey) => {
   if (!report.ticker) return;
   const filter = {
     ticker: report.ticker,
@@ -8,13 +8,12 @@ export const fetchAprData = (report) => {
   };
   return new Promise((resolve, reject) => {
     report.action(filter).then((data) => {
-      console.log('data', data);
-      const formattedData = data.filter(d => d.apr).map(d => ({
+      const formattedData = data.filter(d => d[fieldKey]).map(d => ({
         xaxis1: dayjs(d.date).format('MM/DD/YYY HH:mm:ss'),
-        Price: d.apr,
+        Price: d[fieldKey],
       }));
       if (formattedData.length) {
-        resolve(String(Number(formattedData[formattedData.length-1].Price *100).toFixed(2)) + '%');
+        resolve(widgetFormatter(formattedData[formattedData.length-1]));
       } else {
         resolve('');
       }
