+++ src/components/utils/index.js
@@ -1,11 +1,11 @@
-export function pctFormatter(params) {
-  return Number(params.value*100).toFixed(2) + '%';
+export function pctFormatter(val) {
+  return Number(val*100).toFixed(2) + '%';
 };
 
-export function scoreFormatter(params) {
-  return Number(params.value).toFixed(2);
+export function scoreFormatter(val) {
+  return Number(val).toFixed(2);
 };
 
-export function priceFormat(tickItem) {
-  return String(Number(tickItem*100).toFixed(2))+'%'
+export function numberFormat(val) {
+  return Number(val/1000000).toLocaleString('en-US', {maximumFractionDigits:2})+' M';
 };
