+++ src/pages/Dashboard-terraswap-tlvs/index.js
@@ -5,28 +5,28 @@ import historical from '../../api/v1/historical';
 import { useIsOverflowLockContext } from '../../hooks/useIsOverflowLock';
 import Dashboard from '../../components/Dashboard'
 
-const DashboardTerraSwapApr = () => {
+const DashboardTerraSwapTlvs = () => {
   const reports = [
     {
-      title: 'mNFLX-UST APR',
-      imageUrl: '//whitelist.mirror.finance/images/NFLX.png',
+      title: 'ANC-UST Liquidity',
+      imageUrl: '//whitelist.mirror.finance/images/UST.png',
       value: '',
-      ticker: 'mNFLX-UST',
-      action: historical.getHistoricalCommAprs,
+      ticker: 'ANC-UST',
+      action: historical.getHistoricalTlvs,
     },
     {
-      title: 'mAAPL-UST APR',
+      title: 'mAAPL-UST Liquidity',
       imageUrl: '//whitelist.mirror.finance/images/AAPL.png',
       value: '',
       ticker: 'mAAPL-UST',
-      action: historical.getHistoricalCommAprs,
+      action: historical.getHistoricalTlvs,
     },
     {
-      title: 'mCOIN-UST APR',
+      title: 'mCOIN-UST Liquidity',
       imageUrl: '//whitelist.mirror.finance/images/COIN.png',
       value: '',
       ticker: 'mCOIN-UST',
-      action: historical.getHistoricalCommAprs,
+      action: historical.getHistoricalTlvs,
     },
   ];
 
@@ -50,22 +50,22 @@ const DashboardTerraSwapApr = () => {
   return (
     <Dashboard
       title="Dashboards"
-      breadcrumbItem="MIRROR APRS"
+      breadcrumbItem="POOL LIQUIDITY"
       headerProps={{
-        title: 'TERRASWAP',
+        title: 'Terra Liquidity',
         subTitle: 'Select Pool to Analyze',
-        desc: 'View Historical APR Data',
+        desc: 'View Historical Liquidity Data',
         imgSrc: '//alphadefi.fund/wp-content/uploads/2021/08/687474703a2f2f74657272612e6d6f6e65792f6c6f676f732f74657272615f6c6f676f2e737667.svg'
       }}
       style={styles}
       reports={reports}
       aprTrackers={[AprTrackerShort]}
       widgetFormatter={(data) => {
-        return `${String(Number(data.Price * 100).toFixed(2))}%`;
+        return Number(data.Price).toLocaleString('en-US', {maximumFractionDigits:2});
       }}
-      fieldKey="apr"
+      fieldKey="value"
     />
   );
 }
 
-export default DashboardTerraSwapApr;
+export default DashboardTerraSwapTlvs;
