--- src/pages/Dashboard-tlvs/card-welcome.js
@@ -1,48 +1,0 @@
-import React, { Component } from "react"
-import { Row, Col, Card } from "reactstrap"
-
-//Import Image
-import features from "../../assets/images/crypto/features-img/img-2.png"
-
-class CardWelcome extends Component {
-  render() {
-    return (
-      <React.Fragment>
-        <Card>
-          <div>
-            <Row>
-              <Col lg="9" sm="8">
-                <div className="p-4">
-                  <h5 className="text-primary">Terra Liquidity</h5>
-                  <p>Welcome Alpha Astronaut!</p>
-
-                  <div className="text-muted">
-                    <p className="mb-1">
-                      <i className="mdi mdi-circle-medium align-middle text-primary me-1"/>{" "}
-                      Select Pool to Analyze
-                    </p>
-                    <p className="mb-1">
-                      <i className="mdi mdi-circle-medium align-middle text-primary me-1"/>{" "}
-                      View Historical Liquidity Data
-                    </p>
-                    <p className="mb-0">
-                      <i className="mdi mdi-circle-medium align-middle text-primary me-1"/>{" "}
-                      Locate Trading Opportunities
-                    </p>
-                  </div>
-                </div>
-              </Col>
-              <Col lg="3" sm="4" className="align-self-center">
-                <div>
-                  <img src="//alphadefi.fund/wp-content/uploads/2021/08/687474703a2f2f74657272612e6d6f6e65792f6c6f676f732f74657272615f6c6f676f2e737667.svg" alt="" className="img-fluid d-block" />
-                </div>
-              </Col>
-            </Row>
-          </div>
-        </Card>
-      </React.Fragment>
-    )
-  }
-}
-
-export default CardWelcome
